Limit component.js files to presentation
