+++ packages/components/src/tools-panel/tools-panel-header/hook.js
@@ -7,6 +7,7 @@ import { useMemo } from '@wordpress/element';
  * Internal dependencies
  */
 import * as styles from '../styles';
+import { useToolsPanelContext } from '../context';
 import { useContextSystem } from '../../ui/context';
 import { useCx } from '../../utils/hooks/use-cx';
 
@@ -17,13 +18,17 @@ export function useToolsPanelHeader( props ) {
 	);
 
 	const cx = useCx();
-
 	const classes = useMemo( () => {
 		return cx( styles.ToolsPanelHeader, className );
 	}, [ className ] );
 
+	const { menuItems } = useToolsPanelContext();
+	const hasMenuItems = !! Object.entries( menuItems ).length;
+
 	return {
 		...otherProps,
+		hasMenuItems,
+		menuItems,
 		className: classes,
 	};
 }
